Tidy up tests
