+++ spacy/tests/serialize/test_resource_warning.py
